Release branch only commit: Rework exception handling support
Internal notes for 2015.01-4 to next RC

> Fixed numerous 32bit vs 64bit load/store issues
> Updated the context preserved for longjmp'ing back to bootloader
> Made a yamon specific __uhi_exception implementation
> removed malta32-uboot.ld
> malta32-yamon - match uhi32.ld but with yamon function and no uhi chaining
> Enable the use of offset zero as alternate entry point
> Move exception handler implementations and registration to libhal.a
> Fix EBASE setup to enable use of non kseg0 addresses
> Update gpctx structure to be extensible and store important CP0 state
> Fix BSS setup when the size is not a multiple of 4
> Support compiler generated interrupt handlers
> Fix semi-hosting writes from exception context
> Create overridable symbols for exception handling
> Safely change EBASE with BEV=1
> Use ehb appropriately

External notes for 2013.07-1 to 2014.01-?

== CRT, libuhi.a, libyamon.a, libhal.a ==
Fix yamon versions of UHI to capture fd 0,1,2
Add support for DSP, DSPR2 and MDMX enable bit
Update handling of a0 to fully comply with UHI spec
Move all link base addresses to 0x80200000 (sign extended for 64-bit)
Fix errno handling in UHI functions

+++ libgloss/mips/uhi_get_mem_info.c
@@ -45,9 +45,9 @@
 
 struct s_mem
 {
-  unsigned long size;
-  unsigned long icsize;
-  unsigned long dcsize;
+  unsigned int size;
+  unsigned int icsize;
+  unsigned int dcsize;
 };
 
 /* Size of the memory has been defined .ld file */
@@ -55,7 +55,7 @@ extern char __memory_size[];
 
 void get_mem_info (struct s_mem *mem)
 {
-  mem->size = (unsigned long) __memory_size;
+  mem->size = (unsigned int)(unsigned long) __memory_size;
   return;
 }
